#13 Merge module (comparison/collections → comparison)
