+++ tests/test_comparison.py
